LayoutTests/http/wpt/opener/same-site-child-access-parent-iframe-via-windowproxy.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body onload="addIFrame()">
<script>

let remoteOrigin = new URL(get_host_info()['REMOTE_ORIGIN']);

function addIFrame() {
  let iframeURL = new URL("resources/iframe-open-window.html", document.location);
  iframeURL.protocol = remoteOrigin.protocol;
  iframeURL.hostname = remoteOrigin.hostname;
  iframeURL.port = remoteOrigin.port;

  let openedURL = new URL("resources/send-message-to-opener.html", document.location);

  let iframe = document.createElement('iframe');
  iframe.src = iframeURL + '?location=' + encodeURIComponent(openedURL);
  document.body.appendChild(iframe);
}

const wait_for_message = () => {
  return new Promise((resolve, reject) => {
    window.addEventListener('message', e => {
      resolve(e);
    }, { once: true });
  });
};

promise_test(async function(t) {
  let message = await wait_for_message();
  assert_equals(message.data, 'forwarded: hello world!');

  let domains = await new Promise(resolve => {
    window.testRunner.getAndClearReportedWindowProxyAccessDomains(resolve);
  });
  assert_equals(domains.length, 0);
}, "Test same-site child window accessing parent window's iframe via WindowProxy doesn't log.");
</script>
</body>
